Check if value is numberic
The number validator raises a Liquid error, when the value is a string.
For example:
```liquid
assign object = '{"field1": "one-two-three"}' | parse_json
function c = 'modules/core/lib/validations/number', c: c, object: object, field_name: 'field1', gt: 0
```

Result: `Liquid error: comparison of String with 2147483647 failed`

This change makes an assertion on the type first.
If it's not numeric, then an error is recorded and stops further checks. Otherwise the value is converted to a numeric internal type and comparison checks continues as before.